# swagger.json
{
    "openapi": "3.0.0",
...
    "info": {
        "version": "1.0.0",
        "title": "API de Cortes de Luz",
        "description": "API utilizada para gestionar los datos de los domicilios, empresas y cortes de luz registrados"
    },
    "servers": [
        {
            "url": "http://localhost:2414"
        }
    ],
...
    "paths": {
...
        "/cortes": {
            "get": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Obtener una lista de todos los cortes",
                "description": "Obtiene una lista de cortes con sus respectivos datos",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Agregar un objeto de tipo corte a la lista",
                "description": "Agrega un corte a la lista de cortes",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_corte": {
                                        "type": "number",
                                        "description": "ID del corte",
                                        "example": 0
                                    },
                                    "inicio": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "fin": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cortes/{id_corte}": {
            "get": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Obtener un corte buscandolo por su ID",
                "description": "Obtiene un corte con sus respectivos datos",
                "parameters": [
                    {
                        "name": "id_corte",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Borrar un corte buscandolo por su ID",
                "description": "Borra un corte",
                "parameters": [
                    {
                        "name": "id_corte",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Actualizar todos los datos de un corte",
                "description": "Actualiza los valores de un corte",
                "parameters": [
                    {
                        "name": "id_corte",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_corte": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "inicio": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "fin": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Cortes"
                ],
                "summary": "Actualizar algunos datos de un corte",
                "description": "Actualiza ciertos valores de un corte",
                "parameters": [
                    {
                        "name": "id_corte",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_corte": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "inicio": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "fin": {
                                        "type": "string",
                                        "example": "2024-01-01T07:05:02.000Z"
                                    },
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id_corte": {
                                                "type": "number"
                                            },
                                            "inicio": {
                                                "type": "string"
                                            },
                                            "fin": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cortes/{id_corte}/duracion": {
            "get": {
                "tags": [
                    "Metodos personalizados"
                ],
                "summary": "Busca un corte y consigue su duracion en hs",
                "description": "Busca un corte en la lista de cortes obtiene su inicio y fin para calcular  la duracion del mismo",
                "parameters": [
                    {
                        "name": "id_corte",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "duracion": {
                                                "type": "number",
                                                "example": 0
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cortes/{barrio}/cantidad_de_cortes": {
            "get": {
                "tags": [
                    "Metodos personalizados"
                ],
                "summary": "Busca la cantidad de cortes por barrio",
                "description": "Cuenta la cantidad de cortes por barrio verificando las igualdades",
                "parameters": [
                    {
                        "name": "barrio",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "duracion": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}